requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[aiosqlite]
python-dotenv
pydantic-settings
rich
asgi-correlation-id
python-json-logger
python-jose
python-multipart
bcrypt

aiofiles
b2sdk